requirements.txt
# Multi-agent network requirements
# Based on LangGraph multi-agent example

# Core LangChain packages
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
langchain-tavily>=0.1.0
langchain-experimental>=0.3.0
langchain-core>=0.3.0

# Additional packages for Tavily search
tavily-python>=0.3.0

# LangGraph for multi-agent orchestration
langgraph>=0.2.0

# Data visualization
matplotlib>=3.7.0

# Python REPL for code execution
# Note: This executes code locally and can be unsafe when not sandboxed

# Additional utilities
getpass2>=0.0.3

# Optional: For better development experience
ipython>=8.0.0
jupyter>=1.0.0

# Environment variables and configuration
python-dotenv>=1.0.0

openai>=1.40.0
pandas>=2.2.0
